requirements.txt
# Telegram Bot Dependencies
python-telegram-bot>=22.5,<23
python-dotenv==1.0.0

# Rich for beautiful console output
rich==13.7.1

# HTTP requests
requests==2.31.0

# Optional: For async web scraping
aiohttp==3.13.5

# Web UI and API
Flask==3.1.0
flask-cors==5.0.0
nest-asyncio==1.6.0

# Database
SQLAlchemy==2.0.38

# Data processing
pandas>=3.0.2
numpy>=2.2.6
pydantic==2.10.6
pydantic-settings==2.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
typer>=0.9.0
scikit-learn>=1.3.0

# New LLM Integrations
anthropic>=0.30.0
google-genai>=1.73.1
google-generativeai>=0.7.2
openai>=1.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# NLP Scoring
textblob>=0.17.1

# Vision & Charting
matplotlib>=3.9.0
mplfinance>=0.12.0
pillow==12.2.0

# Technical Analysis
pandas_ta==0.4.71b0
